requirements.txt
PyQt5==5.15.9
sounddevice==0.4.6
numpy==1.24.3
scipy==1.11.2
pyaudiowpatch==0.2.12.8
lameenc==1.5.1
pywin32==306
pyannote.audio